[Feat] 마이페이지 컴포넌트 메타데이터 추가 후 분리
