+++ src/app/mypage/layout.tsx
@@ -1,48 +1,10 @@
-"use client";
-
-import MyPageLoading from "@/components/common/loading/MyPageLoading";
-import MyPageSkeleton from "@/components/common/loading/MyPageSkeleton";
-import UserInfo from "@/components/mypage/UserInfo";
-import AuthGuard from "@/hooks/AuthGuard";
-import { useAuth } from "@/hooks/useAuth";
 import { myPageMetadata } from "@/metadata/mypage/myPageMetadata";
-import { useEffect, useState } from "react";
+import MyPageLayout from "./MyPageLayout";
 
 export const metadata = myPageMetadata;
 
-const MyPageLayout = ({ children }: { children: React.ReactNode }) => {
-  const [, setIsInitialized] = useState(false);
-  const { initializeAuth } = useAuth();
-
-  useEffect(() => {
-    initializeAuth().then(() => setIsInitialized(true));
-  }, [initializeAuth]);
-
-  return (
-    <AuthGuard
-      loadingComponent={
-        <main className="w-full min-h-screen">
-          <div className="max-w-[1320px] m-auto flex flex-col lg:flex-row">
-            <aside className="w-full lg:min-h-screen lg:w-[30%] lg:border-r px-5 pt-14 lg:px-8">
-              <MyPageSkeleton />
-            </aside>
-            <section className="w-full lg:w-[70%]">
-              <MyPageLoading title="" />
-            </section>
-          </div>
-        </main>
-      }
-    >
-      <main className="w-full min-h-screen">
-        <div className="max-w-[1320px] m-auto flex flex-col lg:flex-row">
-          <aside className="w-full lg:min-h-screen lg:w-[30%] lg:border-r px-5 pt-14 lg:px-8">
-            <UserInfo />
-          </aside>
-          <section className="w-full lg:w-[70%]">{children}</section>
-        </div>
-      </main>
-    </AuthGuard>
-  );
+const layout = ({ children }: { children: React.ReactNode }) => {
+  return <MyPageLayout>{children}</MyPageLayout>;
 };
 
-export default MyPageLayout;
+export default layout;
