+++ src/metadata/mypage/myPageMetadata.ts
@@ -1,9 +1,9 @@
 export const myPageMetadata = {
   title: "IVE DIVE",
-  description: "IVE 팬페이지 - 소식페이지",
+  description: "IVE 팬페이지 - 마이페이지",
   openGraph: {
     title: "IVE DIVE",
-    description: "IVE 팬페이지 - 소식페이지",
+    description: "IVE 팬페이지 - 마이페이지",
     images: [
       "https://res.cloudinary.com/dknj7kdek/image/upload/v1737888335/og_nb8ueg.png",
     ],
